requirements.txt
# Interface e Visualização
streamlit==1.31.1
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# Machine Learning
scikit-learn==1.4.0

# LangChain Framework
langchain==0.1.5
langchain-core==0.1.20
langchain-community==0.0.20
langgraph==0.0.20
langsmith==0.0.83

# LLM Providers
openai==1.12.0
langchain-google-genai==0.0.6
langchain-ollama==0.0.1

# Vector Stores
chromadb==0.4.22
faiss-cpu==1.7.4

# Utils
python-dotenv==1.0.1
